Project/daily

See merge request edusoho/edusoho!14231

+++ src/AppBundle/Controller/QuestionBank/QuestionCategoryController.php
@@ -30,6 +30,10 @@ class QuestionCategoryController extends BaseController
 
     public function batchCreateAction(Request $request, $id)
     {
+        if (!$this->getQuestionBankService()->canManageBank($id)) {
+            return $this->createMessageResponse('error', '您不是该题库管理者，不能查看此页面！');
+        }
+
         if ($request->isMethod('POST')) {
             $categoryNames = $request->request->get('categoryNames');
             $categoryNames = trim($categoryNames);
@@ -51,6 +55,10 @@ class QuestionCategoryController extends BaseController
 
     public function editAction(Request $request, $id)
     {
+        if (!$this->getQuestionBankService()->canManageBank($id)) {
+            return $this->createMessageResponse('error', '您不是该题库管理者，不能查看此页面！');
+        }
+
         if ($request->isMethod('POST')) {
             $name = $request->request->get('name', '');
 
@@ -76,6 +84,10 @@ class QuestionCategoryController extends BaseController
 
     public function deleteAction(Request $request, $id)
     {
+        if (!$this->getQuestionBankService()->canManageBank($id)) {
+            return $this->createMessageResponse('error', '您不是该题库管理者，不能查看此页面！');
+        }
+
         $this->getItemCategoryService()->deleteItemCategory($id);
 
         return $this->createJsonResponse(['success' => true]);
